build: Add Pull Request Template (#35)
# Pull Request

## Description


This pull request includes a change to the
`.github/pull_request_template.md` file. The change adds a new template
for pull requests to ensure that contributors provide a summary of the
change, the issue it fixes, relevant motivation and context, and any
dependencies required.

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bR1-R7):
Added a new pull request template to standardize the information
provided by contributors.

Fixes #3