build: Get User's Repositories (#2)
# Pull Request

## Description

This pull request includes several changes to improve the configuration
and repository retrieval functionality. The most important changes
include adding a new dependency, importing necessary modules, creating a
configuration class, and implementing a repository retrieval function.

### Dependency Addition:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711R8):
Added `pygithub` version `2.5.0` to the dependencies.

### Import Statements:
*
[`validator/__main__.py`](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211L6-R17):
Imported `Configuration` and `retrieve_repositories` modules.

### Configuration Class:
*
[`validator/configuration.py`](diffhunk://#diff-d1acdf95b00139fd824c2ec2f961fad04d81299603a1792c38f0a5a93ef53292R1-R32):
Created a `Configuration` class to handle the repository owner and
GitHub token. This class includes methods to initialize the
configuration and retrieve environment variables.

### Repository Retrieval Function:
*
[`validator/repositories.py`](diffhunk://#diff-58cdfc614a2ed6881497d16e1a82c077fed8397d8cd540952e8d3db7864efc97R1-R32):
Implemented the `retrieve_repositories` function to fetch a list of
repositories using the GitHub API. This function logs the count and
names of the retrieved repositories.

Fixes #32