build: Map Security Details to Dataclass (#39)
# Pull Request

## Description

This pull request introduces functionality to check repositories for
required settings and log the results. The main changes include adding a
new repository check function, updating the main application entry point
to use this function, and defining a new data class to represent the
analyzed repositories.

### Main changes:

#### New functionality:
* Added `check_repository` function in `validator/repository_checks.py`
to check repositories for required settings and log the results.
* Defined a new `Repository` data class in `validator/types.py` to
represent the analyzed repository with its settings.

#### Application entry point updates:
* Updated `validator/__main__.py` to use the new `check_repository`
function and log the analyzed repositories.

#### Minor cleanup:
* Removed an unnecessary blank line in `validator/repositories.py`.

Fixes #38