build: Output Analysed Repository to JSON (#41)
# Pull Request

## Description

This pull request introduces changes to the `validator` module to
enhance the handling and storage of analysed repository data. The most
important changes include importing new modules, modifying the `main`
function to use a new data structure, and defining a new `TypedDict` for
structured data representation.

### Enhancements to data handling and storage:

*
[`validator/__main__.py`](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211L9-R11):
Imported `AnalysedRepositories`, `dump`, and `asdict` to support the new
data structure and JSON serialization.
*
[`validator/__main__.py`](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211L18-R26):
Modified the `main` function to use `AnalysedRepositories` for
structured data representation and added code to serialize the analysed
repositories to a JSON file.

### New data structure definition:

*
[`validator/types.py`](diffhunk://#diff-4aad1e39f52c69a8a8d7d31cacc00f823097efefdbbe4b82bb4334ee27372b7bL2-R2):
Imported `TypedDict` to define a new structured data type.
*
[`validator/types.py`](diffhunk://#diff-4aad1e39f52c69a8a8d7d31cacc00f823097efefdbbe4b82bb4334ee27372b7bR13-R17):
Added `AnalysedRepositories` class as a `TypedDict` to represent
analysed repositories with the required settings.

Fixes #40