ci: Check Pull Request Title (#43)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to automate
tasks related to pull requests. The most important change is the
addition of a workflow configuration file that checks the pull request
title for specific prefixes.

New GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR1-R18):
Added a new workflow named "Pull Request Tasks" that triggers on pull
request events (opened, edited, synchronize) and includes a job to check
the pull request title using the `deepakputhraya/action-pr-title`
action. The allowed prefixes for the title are specified in the
configuration.

Fixes #19