ci: Add Dependency Review (#54)
# Pull Request

## Description

This pull request includes a significant update to the GitHub Actions
workflow configuration. The most important change is the addition of a
new job for dependency review.

Workflow configuration updates:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR46-R57):
Added a new job named `dependency-review` which runs on `ubuntu-latest`,
checks out the repository, and uses the
`actions/dependency-review-action` to review dependencies.

Fixes #24