ci: Add Markdown Link Checker (#59)
# Pull Request

## Description

This pull request introduces a new workflow to check Markdown links,
updates the configuration for the link checker, and makes a minor change
to the `README.md` file.

### New Workflow and Configuration:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R42-R59):
Added a new job `check-markdown-links` to validate Markdown links using
the `UmbrellaDocs/action-linkspector` GitHub Action. This job runs on
`ubuntu-latest` and includes steps to checkout the repository and run
the link checker with specified configurations.

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R7):
Added a configuration file for the link checker specifying directories
to include, files to exclude, and status codes to consider as alive.

### Documentation Update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R1):
Updated the title from `#` to `# Repo Standards Validator`.

Fixes #17