ci: Add CodeQL Analysis (#60)
# Pull Request

## Description

This pull request introduces a new CodeQL analysis job to the GitHub
Actions workflow for code checks. The most important change is the
addition of a new job to perform security and quality analysis using
CodeQL.

New job added:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R60-R79):
Added a `run-codeql-analysis` job to perform CodeQL analysis, including
steps for checking out the repository, initializing CodeQL, and
performing the analysis with security and quality queries.

Fixes #23