ci: Check Justfile Format (#61)
# Pull Request

## Description

This pull request introduces a new workflow to check the format of the
`Justfile` in the repository. The changes include adding a new job in
the GitHub Actions workflow and updating the `Justfile` with new
commands for formatting and checking the format.

### Workflow Addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R80-R93):
Added a new job named `check-justfile-format` to the GitHub Actions
workflow. This job sets up Just and runs a format check on the
`Justfile`.

### Justfile Updates:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR1-R11):
Added commands to format the Just code and check for format issues. The
`format` command formats the Just code, and the `format-check` command
checks for format issues.

Fixes #18