docs: Add Code Of Conduct (#68)
# Pull Request

## Description

This pull request introduces a new Code of Conduct for the project. The
document outlines the expected behavior and responsibilities of
community members, as well as the enforcement procedures for violations.

Key additions include:

*
[`docs/CODE_OF_CONDUCT.md`](diffhunk://#diff-5f540ea627655d624f2992755195b9cda3bb919cad13bc3e002cf2bdc520beaeR1-R131):
Added a comprehensive Contributor Covenant Code of Conduct, detailing
community standards, enforcement responsibilities, and guidelines for
addressing violations.

Fixes #6