ci: Add Schema Validation Test Action (#76)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for validating the schema. The most important change is the addition of
the `validate-schema` job in the `.github/workflows/code-checks.yml`
file.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R198-R217):
Added a new job named `validate-schema` to validate the schema using a
series of steps including checking out the repository, setting up
dependencies, downloading the artifact, and running the validation
command.

Fixes #50