chore: Add GitHub Action Summary (#77)
# Pull Request

## Description

This pull request introduces a new feature to generate an action summary
for analysed repositories when running in GitHub Actions. The most
important changes include the addition of the `generate_action_summary`
function and its integration into the main script.

New feature implementation:

*
[`validator/action_summary.py`](diffhunk://#diff-77c11df08149daefb1c1ad1854ac2f65039667e33bae6f9286080c2abbc15650R1-R24):
Added the `generate_action_summary` function to generate an action
summary based on the analysed repositories. This function checks if the
code is running in GitHub Actions and writes the summary to a specified
file.

Integration into main script:

*
[`validator/__main__.py`](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211R12):
Imported the `generate_action_summary` function and called it after
dumping the analysed repositories to a JSON file.
[[1]](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211R12)
[[2]](diffhunk://#diff-ced1c51874c3ff41a0cc0fcbb923f510becf15a3c3b07a78ba5ee9b105eb2211R31)

Fixes #28