ci: Scan GitHub Actions with CodeQL (#94)
# Pull Request

## Description

This pull request includes changes to the
`.github/workflows/code-checks.yml` file to enhance the flexibility and
maintainability of the CI workflow. The most important changes involve
introducing a matrix strategy for language support and updating the
versions of the actions used.

Improvements to CI workflow:

* Introduced a matrix strategy for language support, allowing the
workflow to run for both Python and GitHub Actions.
* Updated the `actions/checkout` action to version `v4.2.2` and renamed
the step for better clarity.
* Updated the `github/codeql-action` versions from `v3.28.0` to
`v3.28.1` for both initialization and analysis steps.
* Modified the `languages` parameter in the CodeQL initialization step
to use the matrix variable `${{ matrix.language }}`.

Fixes #93