feat: Check for Security Policy (#115)
# Pull Request

## Description

This pull request includes several changes to the `validator` module to
add a new feature for checking the presence of a security policy file
(`SECURITY.md`) in repositories. The most important changes include
adding a new function to search for files recursively, updating the
repository check logic to include the security policy check, and adding
corresponding tests.

### New Feature: Security Policy Check

*
[`validator/custom_types.py`](diffhunk://#diff-5a904c05a657819163ec727321af438ab17a521d23010c3d780ff405b14c7adbR15):
Added a new attribute `has_security_policy` to the `Repository` class.
*
[`validator/file.py`](diffhunk://#diff-24c0c8f3e6a102fc3fde014cfadd72d16fa86dc12c9581e48da969e7ab7409eeR1-R20):
Introduced a new function `find_file_recursive` to search for a file in
a directory and its subdirectories.
*
[`validator/repository_checks.py`](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R5):
Updated the `check_repository` function to use `find_file_recursive` for
checking the presence of `SECURITY.md` in repositories.
[[1]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R5)
[[2]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R27-R35)
[[3]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R44)

### Tests

*
[`validator/tests/test_custom_types.py`](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R13):
Added tests to validate the new `has_security_policy` attribute in the
`Repository` class.
[[1]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R13)
[[2]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R25-R28)
*
[`validator/tests/test_file.py`](diffhunk://#diff-9c79888321e180e281b61015178c0272ef9330dfa0f24be320f27857d25d145dR1-R11):
Added a test for the `find_file_recursive` function.

Fixes #109