feat: Check for Code of Conduct (#116)
# Pull Request

## Description

This pull request introduces changes to the repository validation
process by adding a new attribute to check for the presence of a code of
conduct file. The most important changes include updates to the
repository data model, the repository checking function, and the
corresponding test cases.

### Repository validation enhancements:

*
[`validator/custom_types.py`](diffhunk://#diff-5a904c05a657819163ec727321af438ab17a521d23010c3d780ff405b14c7adbR16):
Added a new attribute `has_code_of_conduct` to the `Repository` class to
track the presence of a code of conduct file.

*
[`validator/repository_checks.py`](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R19):
Updated the `check_repository` function to include logic for checking
the presence of a `CODE_OF_CONDUCT.md` file and storing the result in
the `has_code_of_conduct` attribute.
[[1]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R19)
[[2]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84L28-R40)
[[3]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R50)

### Test case updates:

*
[`validator/tests/test_custom_types.py`](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R14):
Modified the `test_repository` function to include assertions for the
new `has_code_of_conduct` attribute to ensure the validation logic is
working correctly.
[[1]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R14)
[[2]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R27)

Fixes #112