feat: Check for Project Technologies (#119)
# Pull Request

## Description

This pull request introduces a new feature to the repository validation
process by adding support for checking the presence of a
`PROJECT_TECHNOLOGIES.md` file. The most important changes include
updates to the `Repository` class, the `check_repository` function, and
corresponding tests.

Updates to repository validation:

*
[`validator/custom_types.py`](diffhunk://#diff-5a904c05a657819163ec727321af438ab17a521d23010c3d780ff405b14c7adbR19):
Added a new boolean field `has_project_technologies` to the `Repository`
class to track the presence of a `PROJECT_TECHNOLOGIES.md` file.

*
[`validator/repository_checks.py`](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R34-R36):
Updated the `check_repository` function to include a check for the
`PROJECT_TECHNOLOGIES.md` file and added the result to the
`AnalysedRepository` dictionary.
[[1]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R34-R36)
[[2]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R46)
[[3]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R59)

Updates to tests:

*
[`validator/tests/test_custom_types.py`](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R17):
Modified the `test_repository` function to include assertions for the
new `has_project_technologies` field.
[[1]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R17)
[[2]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R33)

Fixes #113