feat: Check for License  (#123)
# Pull Request

## Description

This pull request introduces a new feature to validate the presence of a
license file in repositories. The most important changes include updates
to the schema, validation logic, and corresponding tests.

Schema updates:

*
[`tests/schema_validation/repositories_schema.json`](diffhunk://#diff-bbbf42b82f3f8597232147d78cd70cf40ae74122c1f8ef2278bf5cbee23796c5R48-R50):
Added `has_license` field to both the properties and required sections
of the schema.
[[1]](diffhunk://#diff-bbbf42b82f3f8597232147d78cd70cf40ae74122c1f8ef2278bf5cbee23796c5R48-R50)
[[2]](diffhunk://#diff-bbbf42b82f3f8597232147d78cd70cf40ae74122c1f8ef2278bf5cbee23796c5L61-R65)

Validation logic updates:

*
[`validator/custom_types.py`](diffhunk://#diff-5a904c05a657819163ec727321af438ab17a521d23010c3d780ff405b14c7adbR20):
Added `has_license` attribute to the `Repository` class.
*
[`validator/repository_checks.py`](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R37):
Updated `check_repository` function to include logic for checking the
presence of a `LICENSE` file and added `has_license` to the returned
`AnalysedRepository` object.
[[1]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R37)
[[2]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R48)
[[3]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R62)

Test updates:

*
[`validator/tests/test_custom_types.py`](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R18):
Added `has_license` to the test repository setup and assertions in the
`test_repository` function.
[[1]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R18)
[[2]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R35)

Fixes #121