feat: Add Private Vulnerability Disclosures Check (#125)
# Pull Request

## Description

This pull request includes changes to add support for
`private_vulnerability_disclosures` in the repository validation schema
and related components. The most important changes include updates to
the schema, the repository check function, and the test cases.

Updates to schema and types:

*
[`tests/schema_validation/repositories_schema.json`](diffhunk://#diff-bbbf42b82f3f8597232147d78cd70cf40ae74122c1f8ef2278bf5cbee23796c5R34-R36):
Added `private_vulnerability_disclosures` as a new boolean field in the
schema.
*
[`validator/custom_types.py`](diffhunk://#diff-5a904c05a657819163ec727321af438ab17a521d23010c3d780ff405b14c7adbR15):
Added `private_vulnerability_disclosures` as a boolean attribute in the
`Repository` class.

Updates to repository checks:

*
[`validator/repository_checks.py`](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R28):
Updated the `check_repository` function to retrieve and include the
`private_vulnerability_disclosures` status.
[[1]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R28)
[[2]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R44)
[[3]](diffhunk://#diff-a564bb0d5475732c040d805404126747033b4f0829fcc96c7da8d80423860d84R61)

Updates to tests:

*
[`validator/tests/test_custom_types.py`](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R13):
Added test cases to verify the `private_vulnerability_disclosures`
attribute in the `test_repository` function.
[[1]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R13)
[[2]](diffhunk://#diff-62cc5d8a8088e5b85ea04e8a8f4119f63979abd00d6024f6923a2fae5899f4c8R32)

Fixes #126