Add Git Hooks (#151)
# Pull Request

## Description

This pull request includes changes to the `Justfile` and the addition of
new Git hooks to improve the pre-commit and post-commit processes. The
most important changes include adding a `gitleaks-detect` command to the
`Justfile`, creating a new `post-commit` hook, and enhancing the
`pre-commit` hook with multiple checks.

Enhancements to `Justfile`:

* Added a `gitleaks-detect` command to detect secrets in the codebase.

New Git hooks:

* Created a new `post-commit` hook to update the Git index.
* Enhanced the `pre-commit` hook to run `gitleaks-detect`,
`prettier-check`, and `format-check` commands, and provide appropriate
feedback based on the results.

Fixes #10