chore: Update SonarCloud and Configuration (#160)
# Pull Request

## Description

This pull request includes changes to the
`.github/workflows/code-checks.yml` file to update the SonarCloud scan
action and remove an unnecessary step.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L238-R239):
Updated the SonarCloud scan action from version `v4.2.1` to `v5.0.0`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L238-R239):
Removed the step to override the coverage source path for SonarCloud as
it is no longer necessary.