chore: Update SuperLinter EditorConfig Checker (#165)
# Pull Request

## Description

This pull request includes changes to configuration files for the
super-linter and code checks. The most important changes involve
renaming a configuration file and updating a workflow file to reflect
this change.

Changes to configuration files:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L5-R5):
Renamed from `.ecrc` and corrected the capitalization of
`SpacesAfterTabs`.

Updates to workflow file:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R36):
Added `EDITORCONFIG_FILE_NAME` to the environment variables to reflect
the new configuration file name.