chore: Remove Pinact Configuration (#189)
# Pull Request

## Description

This pull request removes the configuration file for `pinact` and
updates related commands in the `Justfile` to no longer reference the
deleted file. The most important changes are as follows:

### Removal of `pinact` configuration:

*
[`.github/other-configurations/pinact.yml`](diffhunk://#diff-c2e684c61036562879b1062fb69ab67d446e19354f7a45ef92b14f611a051bc5L1-L9):
Deleted the `pinact.yml` configuration file, which included schema
references, versioning, and ignored actions.

### Updates to `Justfile` commands:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL162-R170):
Modified `pinact`-related commands (`pinact-run`, `pinact-check`, and
`pinact-update`) to remove the `-c
.github/other-configurations/pinact.yml` flag, as the configuration file
is no longer used.