chore: Remove `.editorconfig-checker.json` (#207)
# Pull Request

## Description

This pull request removes unused configurations and updates a dependency
version. The most important changes include the removal of the
`.editorconfig-checker.json` file and its related references, as well as
a minor version update for `lefthook`.

### Removal of unused configurations:
*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Deleted the file entirely as it is no longer in use.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable, which
referenced the deleted `.editorconfig-checker.json` file.

### Dependency updates:
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of `lefthook` from `1.11.11` to `1.11.13` to
ensure compatibility or include minor improvements.