ci: Add Output Testing to Local Action Job (#223)
# Pull Request

## Description

This pull request introduces a new test step in the CI workflow and
updates a test script to align with a renamed file. These changes
enhance the reliability of the testing process and ensure consistency
with updated file naming.

### Workflow Updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R133-R134):
Added a new step named "Test Output" to run the `tests/test_output.sh`
script as part of the CI pipeline.

### Test Script Updates:
*
[`tests/test_output.sh`](diffhunk://#diff-8147574e4cf9f36dc56066ea4e332658da4d0b612512eb5e18a85496244e85aeL5-R8):
Updated the file checks to reference `repositories.json` instead of the
outdated `repository_statistics.json`, ensuring the script aligns with
the renamed file.