ci: Add release package workflow (#226)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow for releasing
packages and makes several updates to rename the "analyser" to
"validator" throughout the codebase. Below is a summary of the most
important changes:

### New GitHub Actions Workflow

* Added `.github/workflows/release-package.yml` to automate the process
of building, pushing, and testing Docker images upon tagging a new
version. This includes steps for logging into the container registry,
extracting metadata, building and pushing the Docker image, generating
artifact attestations, and running tests.

### Renaming "Analyser" to "Validator"

* Updated comments in `Justfile` to reflect the renaming, changing
"analyser" to "validator" in the `clean-generated-files` section.
* Updated a comment in `run.sh` to rename "analyser" to "validator" and
ensured the script runs the `validator` module.
* Modified the docstring in `validator/configuration.py` to rename
"analyser" to "validator" in the `Configuration` class description.